@@ -23,7 +23,7 @@ public class FcmPacketExtension implements ExtensionElement {
   @Override
   public String toXML(String enclosingNamespace) {
     // TODO: 1. Do we need to scape the json? StringUtils.escapeForXML(json) 2. How to use the enclosing namespace?
-    return String.format("<%s xmlns=\"%s\">%s</%s>", getElementName(), getNamespace(), json, Util.FCM_ELEMENT_NAME);
+    return "<%s xmlns=\"%s\">%s</%s>".formatted(getElementName(), getNamespace(), json, Util.FCM_ELEMENT_NAME);
   }
 
   public Stanza toPacket() {
