@@ -41,14 +41,15 @@ public class App {
 
     public static void main(String[] args) throws IOException {
 
-        System.out.println("               _ _____                _       \n" +
-                "              | |  __ \\              | |      \n" +
-                "  _ __ ___  __| | |__) |_ _ _ __   __| | __ _ \n" +
-                " | '__/ _ \\/ _` |  ___/ _` | '_ \\ / _` |/ _` |\n" +
-                " | | |  __/ (_| | |  | (_| | | | | (_| | (_| |\n" +
-                " |_|  \\___|\\__,_|_|   \\__,_|_| |_|\\__,_|\\__,_|               - p2p chat, https://github.com/redPanda-project/redpandaj\n" +
-                "                                              \n" +
-                "                                              ");
+        System.out.println("""
+                               _ _____                _      \s
+                              | |  __ \\              | |     \s
+                  _ __ ___  __| | |__) |_ _ _ __   __| | __ _\s
+                 | '__/ _ \\/ _` |  ___/ _` | '_ \\ / _` |/ _` |
+                 | | |  __/ (_| | |  | (_| | | | | (_| | (_| |
+                 |_|  \\___|\\__,_|_|   \\__,_|_| |_|\\__,_|\\__,_|               - p2p chat, https://github.com/redPanda-project/redpandaj
+                                                             \s
+                                                              """);
 
         System.out.println("Starting redpanda " + App.class.getPackage().getImplementationVersion());
 
