@@ -88,8 +88,7 @@ public class Utils {
     public static String formatDuration(Duration duration) {
         long seconds = duration.getSeconds();
         long absSeconds = Math.abs(seconds);
-        String positive = String.format(
-                "%d:%02d:%02d",
+        String positive = "%d:%02d:%02d".formatted(
                 absSeconds / 3600,
                 (absSeconds % 3600) / 60,
                 absSeconds % 60);
